PythonServer/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
gunicorn
mysql-connector-python
python-dotenv